requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
pdfplumber==0.10.0
Pillow==10.0.0
pytesseract==0.3.10
langchain==0.0.350
langchain-community==0.0.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
gunicorn==21.2.0
requests==2.31.0